authentication file updated
